+++ CHANGELOG.md
@@ -234,3 +234,12 @@
 
 ## 0.9.1
 - mainField and otherField on DataBean.join are optional
+
+# 0.10.0
+- renamed PropertyCompareType to CompareType
+- refactored CompareType values to lower camel case
+- added Expression
+  - added Expression convenience methods to create Filter objects
+  - added Expression convenience methods to create Sort objects
+- PropertyCompare has been replaced by CompareFilter
+- FieldSort has been replaced by ExpressionSort